pkg/endpoint: move API model-related code into separate file
The file endpoint.go is massive. Put endpoint-API-model-related code into its own file.

Signed-off by: Ian Vernon <ian@cilium.io>
